app-fast-api-ood/requirements.txt
uvicorn~=0.32.1
fastapi~=0.115.5
pillow~=11.0.0
numpy~=2.1.3
matplotlib~=3.9.2
ultralytics~=8.3.36